cleaning up extinction in tributaries by adding tributary_dust_add
